+++ tests/unit/Video/Info/VideoStreamTest.php
@@ -76,4 +76,58 @@ class VideoStreamTest extends TestCase
         $stream = new VideoStream($data);
         self::assertNull($stream->isAvc());
     }
+
+    public function testGetFps(): void
+    {
+        $data = $this->getExampleFFProbeData()['streams'][0];
+        $d    = array_merge($data, [
+            'r_frame_rate' => '25/1'
+        ]);
+        self::assertEquals(25, (new VideoStream($d))->getFps());
+
+        $d = array_merge($data, [
+            'r_frame_rate' => '24000/1001'
+        ]);
+        self::assertEquals(23.98, (new VideoStream($d))->getFps(2));
+
+        $d = array_merge($data, [
+            'r_frame_rate' => '24000/1001'
+        ]);
+        self::assertEquals(24, (new VideoStream($d))->getFps(0));
+
+        $d = array_merge($data, [
+            'r_frame_rate' => null,
+            'duration'     => 1,
+            'nb_frames'    => 30
+        ]);
+        self::assertEquals(30, (new VideoStream($d))->getFps(0));
+
+        $d = array_merge($data, [
+            'r_frame_rate' => null,
+            'duration'     => 1000,
+            'nb_frames'    => 60002
+        ]);
+        self::assertEquals(60.002, (new VideoStream($d))->getFps(3));
+
+        $d = array_merge($data, [
+            'r_frame_rate' => 'A/a',
+            'duration'     => 1000,
+            'nb_frames'    => 60002
+        ]);
+        self::assertEquals(60.002, (new VideoStream($d))->getFps(3));
+
+        $d = array_merge($data, [
+            'r_frame_rate' => null,
+            'duration'     => null,
+            'nb_frames'    => 60002
+        ]);
+        self::assertNull((new VideoStream($d))->getFps(1));
+
+        $d = array_merge($data, [
+            'r_frame_rate' => null,
+            'duration'     => null,
+            'nb_frames'    => null
+        ]);
+        self::assertNull((new VideoStream($d))->getFps(1));
+    }
 }